Add switch between widgets configuration
