+++ README.md
@@ -5,7 +5,7 @@
 
 Airpods Mac OS App which show you the battery percentage of your airpods and Airpods Pro from the status bar or the associated widget
 
-![Image of AirPods Widget](/images/widget_demo_small.png)
+![Image of AirPods Widget](airpods-battery-widget.png)
 ![Airpods Max Battery Widget](/images/airpodsMax-widget-demo.png)
 
 A small Mac OS App that allow you to see easily your AirPods Battery (Case/ Left / Right) in real-time !. It is a shortcut to remove the long and painful access from the Bluetooth Tab that Mac OS X provide.
